refactor: Program.cs aufteilen, DTOs validieren, Tests-Projekt, SETUP.md
- Program.cs von 1300 auf ~160 Zeilen reduziert
- 10 Endpoint-Klassen in Extensions/ erstellt
- AppDbContext in Data/ ausgelagert, ProgressConst in Models/
- Alle 17 DTOs mit DataAnnotations ausgestattet
- xUnit-Testprojekt ClimbConnect.Tests angelegt
- docs/SETUP.md vollständig dokumentiert
