+++ backend/ClimbConnect.API/Dtos/AppointmentSubscribeDto.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record AppointmentSubscribeDto(string? Comment);
+public record AppointmentSubscribeDto([MaxLength(500)] string? Comment);
