+++ backend/ClimbConnect.API/Dtos/AreaCreateDto.cs
@@ -1,3 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record AreaCreateDto(string Name, string? Location, string? Description, string? ImageUrl);
+public record AreaCreateDto(
+    [Required] [StringLength(200, MinimumLength = 1)] string  Name,
+    [MaxLength(200)]                                  string? Location,
+    [MaxLength(2000)]                                 string? Description,
+    [MaxLength(500)]                                  string? ImageUrl
+);
