+++ backend/ClimbConnect.API/Dtos/LoginDto.cs
@@ -1,3 +1,8 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record LoginDto(string Email, string Password);
+public record LoginDto(
+    [Required] [EmailAddress] [MaxLength(200)] string Email,
+    [Required] [MaxLength(100)]                string Password
+);
