+++ backend/ClimbConnect.API/Dtos/ProgressCreateDto.cs
@@ -1,12 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
 public record ProgressCreateDto(
-    int RouteId,
-    string Status,
-    string ClimbingStyle,
-    int Attempts,
-    string? Notes,
-    DateOnly Date,
-    string? SubjectiveGrade,
-    string? SubjectiveGradeComment
+    [Required] [Range(1, int.MaxValue)] int     RouteId,
+    [Required] [MaxLength(20)]          string  Status,
+    [Required] [MaxLength(20)]          string  ClimbingStyle,
+    [Range(1, 9999)]                    int     Attempts,
+    [MaxLength(1000)]                   string? Notes,
+    [Required]                          DateOnly Date,
+    [MaxLength(20)]                     string?  SubjectiveGrade,
+    [MaxLength(500)]                    string?  SubjectiveGradeComment
 );
