+++ backend/ClimbConnect.API/Dtos/RegisterDto.cs
@@ -1,3 +1,9 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record RegisterDto(string Username, string Email, string Password);
+public record RegisterDto(
+    [Required] [StringLength(50, MinimumLength = 3)] string Username,
+    [Required] [EmailAddress] [StringLength(200)]    string Email,
+    [Required] [MinLength(8)] [MaxLength(100)]        string Password
+);
