+++ backend/ClimbConnect.API/Dtos/ReportStatusUpdateDto.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record ReportStatusUpdateDto(string Status);
+public record ReportStatusUpdateDto([Required] [MaxLength(20)] string Status);
