+++ backend/ClimbConnect.API/Dtos/RouteUpdateDto.cs
@@ -1,9 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
 public record RouteUpdateDto(
-    string Name,
-    string? Grade,
-    int? LengthMeters,
-    string? Style,
-    string? Description
+    [Required] [StringLength(200, MinimumLength = 1)] string  Name,
+    [MaxLength(20)]                                   string? Grade,
+    [Range(1, 3000)]                                  int?    LengthMeters,
+    [MaxLength(100)]                                  string? Style,
+    [MaxLength(2000)]                                 string? Description
 );
