+++ backend/ClimbConnect.API/Dtos/SectorCreateDto.cs
@@ -1,3 +1,8 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record SectorCreateDto(string Name, string? Description);
+public record SectorCreateDto(
+    [Required] [StringLength(200, MinimumLength = 1)] string  Name,
+    [MaxLength(2000)]                                 string? Description
+);
