+++ backend/ClimbConnect.API/Dtos/SectorUpdateDto.cs
@@ -1,3 +1,8 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record SectorUpdateDto(string Name, string? Description);
+public record SectorUpdateDto(
+    [Required] [StringLength(200, MinimumLength = 1)] string  Name,
+    [MaxLength(2000)]                                 string? Description
+);
