+++ backend/ClimbConnect.API/Dtos/UserProfileUpdateDto.cs
@@ -1,3 +1,8 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace ClimbConnect.API.Dtos;
 
-public record UserProfileUpdateDto(string? Bio, string? PreferredGradeScale);
+public record UserProfileUpdateDto(
+    [MaxLength(500)] string? Bio,
+    [MaxLength(20)]  string? PreferredGradeScale
+);
